Modification de la barre de navigation de jeu
### Liens de la page jouer.php
- Réparation des liens brisés "Classement" et "FAQ"
- Suppression des liens vers les discords camp

### Structure
- mise en place d'une base de refactorisation
